api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
minio
python-multipart
pydantic-settings
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
